Create GasStats component
