+++ src/main.ts
@@ -36,6 +36,7 @@ Vue.use(VTooltip);
 Vue.component("Header", () => import("@/components/Header.vue"));
 Vue.component("Footer", () => import("@/components/Footer.vue"));
 Vue.component("Card", () => import("@/components/Card.vue"));
+Vue.component("GasStats", () => import("@/components/ugas/GasStats.vue"));
 Vue.component("CardLink", () => import("@/components/CardLink.vue"));
 Vue.component("Button", () => import("@/components/Button.vue"));
 Vue.component("chart", ECharts);
